Remove kubernetes-deploy* (#622)
* Remove kubernetes-deploy*
* Rename class
* Remove template_dir and only_filenames from RenderTask
* Remove template_dir from DeployTask
* Remove allow_globals from DeployTask
* Move unsafe parallel tests to the serial suite
* Remove duplicated resources from pruning whitelis
