+++ 1.0-Upgrade.md
@@ -15,7 +15,14 @@ custom resources. See [blacklist](https://github.com/Shopify/kubernetes-deploy/b
 
 ## Public API changes
 
-The only breaking change in the public API (so far) is the renaming of the `KubernetesDeploy` namespace to `Krane`. Otherwise, the APIs of the major public classes (`DeployTask`, `RenderTask`, `RunnerTask`, and `RestartTask`) have not changed between 0.31.1 and 1.0.0. If you're curious about this API, check the comment-based docs on these classes or the [rendered documentation at RubyGems.org](https://www.rubydoc.info/gems/kubernetes-deploy/1.0.0/KubernetesDeploy/DeployTask).
+Important changes:
+- Removed `KubernetesDeploy#*`: use the appropriate `Krane::*` class instead.
+- Removed `template_dir` from `DeployTask#initialize`: use `template_paths` instead.
+- Removed `allow_protected_ns` from `DeployTask#run`: use `protected_namespaces` instead.
+- Removed `template_dir` from `RenderTask#initialize` and `only_filenames` from `RenderTask#run`: use `template_paths` instead.
+- Removed `allow_globals` from `DeployTask`: use `GlobalDeployTask` instead.
+
+If you want to see the current state of this API, check the comment-based docs on these classes or the [rendered documentation at RubyGems.org](https://www.rubydoc.info/gems/kubernetes-deploy/1.0.0/KubernetesDeploy/DeployTask).
 
 ## Command-line interface changes
