--- exe/kubernetes-restart
@@ -1,33 +1,0 @@
-#!/usr/bin/env ruby
-# frozen_string_literal: true
-
-require 'optparse'
-
-require 'kubernetes-deploy/restart_task'
-require 'krane/options_helper'
-require 'krane/label_selector'
-
-raw_deployments = nil
-max_watch_seconds = nil
-selector = nil
-ARGV.options do |opts|
-  opts.on("--deployments=LIST") { |v| raw_deployments = v.split(",") }
-  opts.on("--max-watch-seconds=seconds") { |t| max_watch_seconds = t.to_i }
-  opts.on("--selector=SELECTOR", "Restarts deployments matching selector (format: k1=v1,k2=v2)") do |s|
-    selector = Krane::LabelSelector.parse(s)
-  end
-  opts.parse!
-end
-
-namespace = ARGV[0]
-context = ARGV[1]
-
-restart = KubernetesDeploy::RestartTask.new(namespace: namespace, context: context,
-   max_watch_seconds: max_watch_seconds)
-begin
-  restart.run!(raw_deployments, selector: selector)
-rescue KubernetesDeploy::DeploymentTimeoutError
-  exit(70)
-rescue KubernetesDeploy::FatalDeploymentError
-  exit(1)
-end
