+++ lib/krane/cluster_resource_discovery.rb
@@ -20,7 +20,7 @@ module Krane
       black_list = %w(Namespace Node ControllerRevision)
       api_versions = fetch_api_versions
 
-      fetch_resources(namespaced: namespaced).map do |resource|
+      fetch_resources(namespaced: namespaced).uniq { |r| r['kind'] }.map do |resource|
         next unless resource['verbs'].one? { |v| v == "delete" }
         next if black_list.include?(resource['kind'])
         group_versions = api_versions[resource['apigroup'].to_s]
