+++ lib/krane/render_task.rb
@@ -12,12 +12,10 @@ module Krane
     #
     # @param logger [Object] Logger object (defaults to an instance of Krane::FormattedLogger)
     # @param current_sha [String] The SHA of the commit
-    # @param template_dir [String] Path to a directory with templates to render (deprecated)
     # @param template_paths [Array<String>] An array of template paths to render
     # @param bindings [Hash] Bindings parsed by Krane::BindingsParser
-    def initialize(logger: nil, current_sha:, template_dir: nil, template_paths: [], bindings:)
+    def initialize(logger: nil, current_sha:, template_paths: [], bindings:)
       @logger = logger || Krane::FormattedLogger.build
-      @template_dir = template_dir
       @template_paths = template_paths.map { |path| File.expand_path(path) }
       @bindings = bindings
       @current_sha = current_sha
@@ -36,16 +34,15 @@ module Krane
     # Runs the task, raising exceptions in case of issues
     #
     # @param stream [IO] Place to stream the output to
-    # @param only_filenames [Array<String>] List of filenames to render
     #
     # @return [nil]
-    def run!(stream, only_filenames = [])
+    def run!(stream)
       @logger.reset
       @logger.phase_heading("Initializing render task")
 
-      ts = TemplateSets.from_dirs_and_files(paths: template_sets_paths(only_filenames), logger: @logger)
+      ts = TemplateSets.from_dirs_and_files(paths: @template_paths, logger: @logger)
 
-      validate_configuration(ts, only_filenames)
+      validate_configuration(ts)
       count = render_templates(stream, ts)
 
       @logger.summary.add_action("Successfully rendered #{count} template(s)")
@@ -56,20 +53,6 @@ module Krane
     end
 
     private
-
-    def template_sets_paths(only_filenames)
-      if @template_paths.present?
-        # Validation will catch @template_paths & @template_dir being present
-        @template_paths
-      elsif only_filenames.blank?
-        [File.expand_path(@template_dir || '')]
-      else
-        absolute_template_dir = File.expand_path(@template_dir || '')
-        only_filenames.map do |name|
-          File.join(absolute_template_dir, name)
-        end
-      end
-    end
 
     def render_templates(stream, template_sets)
       @logger.phase_heading("Rendering template(s)")
@@ -102,30 +85,12 @@ module Krane
       raise InvalidTemplateError.new("Template is not valid YAML. #{exception.message}", filename: filename)
     end
 
-    def validate_configuration(template_sets, filenames)
+    def validate_configuration(template_sets)
       @logger.info("Validating configuration")
       errors = []
-      if @template_dir.present? && @template_paths.present?
-        errors << "template_dir and template_paths can not be combined"
-      elsif @template_dir.blank? && @template_paths.blank?
-        errors << "template_dir or template_paths must be set"
+      if @template_paths.blank?
+        errors << "template_paths must be set"
       end
-
-      if filenames.present?
-        if @template_dir.nil?
-          errors << "template_dir must be set to use filenames"
-        else
-          absolute_template_dir = File.expand_path(@template_dir)
-          filenames.each do |filename|
-            absolute_file = File.expand_path(File.join(@template_dir, filename))
-            unless absolute_file.start_with?(absolute_template_dir)
-              errors << "Filename \"#{absolute_file}\" is outside the template directory," \
-              " which was resolved as #{absolute_template_dir}"
-            end
-          end
-        end
-      end
-
       errors += template_sets.validate
 
       unless errors.empty?
