--- lib/kubernetes-deploy.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-require 'kubernetes-deploy/deploy_task'
-require 'kubernetes-deploy/render_task'
-require 'kubernetes-deploy/restart_task'
-require 'kubernetes-deploy/runner_task'
