+++ test/fixtures/crd/things.yml
@@ -2,6 +2,8 @@ apiVersion: apiextensions.k8s.io/v1beta1
 kind: CustomResourceDefinition
 metadata:
   name: things.stable.example.io
+  labels:
+    app: krane
 spec:
   group: stable.example.io
   version: v1
