+++ test/fixtures/crd/widgets.yml
@@ -2,6 +2,8 @@ apiVersion: apiextensions.k8s.io/v1beta1
 kind: CustomResourceDefinition
 metadata:
   name: widgets.stable.example.io
+  labels:
+    app: krane
   annotations:
     krane.shopify.io/prunable: "true"
 spec:
