+++ test/fixtures/crd/widgets_deprecated.yml
@@ -2,6 +2,8 @@ apiVersion: apiextensions.k8s.io/v1beta1
 kind: CustomResourceDefinition
 metadata:
   name: widgets.stable.example.io
+  labels:
+    app: krane
   annotations:
     kubernetes-deploy.shopify.io/prunable: "true"
 spec:
