+++ test/fixtures/crd/with_custom_conditions.yml
@@ -2,6 +2,8 @@ apiVersion: apiextensions.k8s.io/v1beta1
 kind: CustomResourceDefinition
 metadata:
   name: customizeds.stable.example.io
+  labels:
+    app: krane
   annotations:
     krane.shopify.io/instance-rollout-conditions: '{
       "success_conditions": [
