+++ test/fixtures/crd/with_default_conditions.yml
@@ -2,6 +2,8 @@ apiVersion: apiextensions.k8s.io/v1beta1
 kind: CustomResourceDefinition
 metadata:
   name: parameterizeds.stable.example.io
+  labels:
+    app: krane
   annotations:
     krane.shopify.io/instance-rollout-conditions: "true"
 spec:
