+++ test/fixtures/pvc/wait_for_first_consumer_storage_class.yml
@@ -2,5 +2,7 @@ apiVersion: storage.k8s.io/v1
 kind: StorageClass
 metadata:
   name: k8s-deploy-test
+  labels:
+    app: krane
 provisioner: kubernetes.io/no-provisioner
 volumeBindingMode: WaitForFirstConsumer
