+++ test/integration/krane_deploy_test.rb
@@ -108,7 +108,7 @@ class KraneDeployTest < Krane::IntegrationTest
       prune_matcher("statefulset", "apps", "stateful-busybox"),
       prune_matcher("job", "batch", "hello-job"),
       prune_matcher("poddisruptionbudget", "policy", "test"),
-      prune_matcher("networkpolicy", "networking.k8s.io", "allow-all-network-policy"),
+      prune_matcher("networkpolicy", "*", "allow-all-network-policy"),
       prune_matcher("secret", "", "hello-secret"),
       prune_matcher("replicaset", "apps", "bare-replica-set"),
       prune_matcher("serviceaccount", "", "build-robot"),
@@ -192,7 +192,7 @@ class KraneDeployTest < Krane::IntegrationTest
   def test_refuses_deploy_to_protected_namespace_with_override_if_pruning_enabled
     generated_ns = @namespace
     @namespace = 'default'
-    assert_deploy_failure(deploy_fixtures("hello-cloud", allow_protected_ns: true, prune: true))
+    assert_deploy_failure(deploy_fixtures("hello-cloud", prune: true))
     assert_logs_match_all([
       "Configuration invalid",
       "- Refusing to deploy to protected namespace 'default' with pruning enabled",
@@ -1520,114 +1520,6 @@ unknown field \"myKey\" in io.k8s.apimachinery.pkg.apis.meta.v1.ObjectMeta",
     refute_logs_match("YWRtaW4=")
   end
 
-  # Note: These tests assume a default storage class with a dynamic provisioner and 'Immediate' bind
-  def test_pvc
-    pvname = "local0001"
-    storage_class_name = "k8s-deploy-test"
-
-    assert_deploy_success(deploy_fixtures("pvc", subset: ["wait_for_first_consumer_storage_class.yml"]))
-
-    TestProvisioner.prepare_pv(pvname, storage_class_name: storage_class_name)
-    assert_deploy_success(deploy_fixtures("pvc"))
-
-    assert_logs_match_all([
-      "Successfully deployed 4 resource",
-      "Successful resources",
-      %r{PersistentVolumeClaim/with-storage-class\s+Bound},
-      %r{PersistentVolumeClaim/without-storage-class\s+Bound},
-      %r{Pod/pvc\s+Succeeded},
-      %r{StorageClass/k8s-deploy-test\s+Exists},
-    ], in_order: true)
-
-  ensure
-    kubeclient.delete_persistent_volume(pvname)
-    storage_v1_kubeclient.delete_storage_class(storage_class_name)
-  end
-
-  def test_pvc_no_bind
-    pvname = "local0002"
-    storage_class_name = "k8s-deploy-test-no-bind"
-
-    result = deploy_fixtures("pvc", subset: ["wait_for_first_consumer_storage_class.yml"]) do |fixtures|
-      sc = fixtures["wait_for_first_consumer_storage_class.yml"]["StorageClass"].first
-      sc["metadata"]["name"] = storage_class_name
-    end
-    assert_deploy_success(result)
-
-    TestProvisioner.prepare_pv(pvname, storage_class_name: storage_class_name)
-    result = deploy_fixtures("pvc", subset: ["pvc.yml"]) do |fixtures|
-      pvc = fixtures["pvc.yml"]["PersistentVolumeClaim"].first
-      pvc["spec"]["storageClassName"] = storage_class_name
-    end
-    assert_deploy_success(result)
-
-    assert_logs_match_all([
-      "Successfully deployed 2 resource",
-      "Successful resources",
-      %r{PersistentVolumeClaim/with-storage-class\s+Pending},
-      %r{PersistentVolumeClaim/without-storage-class\s+Bound},
-    ], in_order: true)
-
-  ensure
-    kubeclient.delete_persistent_volume(pvname)
-    storage_v1_kubeclient.delete_storage_class(storage_class_name)
-  end
-
-  def test_pvc_immediate_bind
-    pvname = "local0003"
-    storage_class_name = "k8s-deploy-test-immediate-bind"
-
-    result = deploy_fixtures("pvc", subset: ["wait_for_first_consumer_storage_class.yml"]) do |fixtures|
-      sc = fixtures["wait_for_first_consumer_storage_class.yml"]["StorageClass"].first
-      sc["metadata"]["name"] = storage_class_name
-      sc["volumeBindingMode"] = "Immediate"
-    end
-    assert_deploy_success(result)
-    TestProvisioner.prepare_pv(pvname, storage_class_name: storage_class_name)
-    result = deploy_fixtures("pvc", subset: ["pvc.yml"]) do |fixtures|
-      pvc = fixtures["pvc.yml"]["PersistentVolumeClaim"].first
-      pvc["spec"]["storageClassName"] = storage_class_name
-    end
-    assert_deploy_success(result)
-
-    assert_logs_match_all([
-      "Successfully deployed 2 resource",
-      "Successful resources",
-      %r{PersistentVolumeClaim/with-storage-class\s+Bound},
-      %r{PersistentVolumeClaim/without-storage-class\s+Bound},
-    ], in_order: true)
-
-  ensure
-    kubeclient.delete_persistent_volume(pvname)
-    storage_v1_kubeclient.delete_storage_class(storage_class_name)
-  end
-
-  def test_pvc_no_pv
-    storage_class_name = "k8s-deploy-test-no-pv"
-
-    result = deploy_fixtures("pvc", subset: ["wait_for_first_consumer_storage_class.yml"]) do |fixtures|
-      sc = fixtures["wait_for_first_consumer_storage_class.yml"]["StorageClass"].first
-      sc["metadata"]["name"] = storage_class_name
-    end
-    assert_deploy_success(result)
-
-    result = deploy_fixtures("pvc", subset: ["pvc.yml", "pod.yml"]) do |fixtures|
-      pvc = fixtures["pvc.yml"]["PersistentVolumeClaim"].first
-      pvc["spec"]["storageClassName"] = storage_class_name
-    end
-    assert_deploy_failure(result)
-
-    assert_logs_match_all([
-      "Failed to deploy 1 priority resource",
-      "Pod/pvc: TIMED OUT (timeout: 10s)",
-      %r{Pod could not be scheduled because 0/\d+ nodes are available:},
-      /\d+ node[(]s[)] didn't find available persistent volumes to bind./,
-    ], in_order: true)
-
-  ensure
-    storage_v1_kubeclient.delete_storage_class(storage_class_name)
-  end
-
   def test_pvc_no_pv_or_sc
     storage_class_name = "k8s-deploy-test-no-pv-or-sc"
 
@@ -1671,18 +1563,8 @@ unknown field \"myKey\" in io.k8s.apimachinery.pkg.apis.meta.v1.ObjectMeta",
     ], in_order: true)
   end
 
-  def test_deploy_allow_globals_warns
+  def test_deploy_globals_fails
     result = deploy_fixtures("globals")
-    assert_deploy_success(result)
-    assert_logs_match_all([
-      "The ability for this task to deploy global resources will be removed in the next version,"\
-      " which will affect the following resources:",
-      '    testing-storage-class (StorageClass) in ',
-    ], in_order: true)
-  end
-
-  def test_deploy_globals_without_allow_globals_fails
-    result = deploy_fixtures("globals", allow_globals: false)
     assert_deploy_failure(result)
     assert_logs_match_all([
       'This command is namespaced and cannot be used to deploy global resources.',
