+++ test/unit/cluster_resource_discovery_test.rb
@@ -32,7 +32,7 @@ class ClusterResourceDiscoveryTest < Krane::TestCase
     crd = mocked_cluster_resource_discovery(api_resources_not_namespaced_full_response)
     kinds = crd.prunable_resources(namespaced: false)
 
-    assert_equal(kinds.length, 13)
+    assert_equal(kinds.length, 12)
     %w(PriorityClass StorageClass).each do |expected_kind|
       assert kinds.one? { |k| k.include?(expected_kind) }
     end
@@ -47,7 +47,7 @@ class ClusterResourceDiscoveryTest < Krane::TestCase
     crd = mocked_cluster_resource_discovery(api_resources_namespaced_full_response, namespaced: true)
     kinds = crd.prunable_resources(namespaced: true)
 
-    assert_equal(kinds.length, 27)
+    assert_equal(kinds.length, 25)
     %w(ConfigMap CronJob Deployment).each do |expected_kind|
       assert kinds.one? { |k| k.include?(expected_kind) }
     end
